requirements.txt
streamlit
pdfplumber
openai
python-dotenv
babel
altair
xhtml2pdf
pandas
